requirements.txt
pandas
gunicorn
pillow>=8.0
face_recognition
Flask
pytest
mockito